+++ spigot/src/main/kotlin/com/dreamdisplays/commands/subcommands/HelpCommand.kt
@@ -9,6 +9,7 @@ class HelpCommand : SubCommand {
 
     override val name = "help"
     override val permission = config.permissions.help
+    override val playerOnly = true
 
     override fun execute(sender: CommandSender, args: Array<String?>) {
         val player = (sender as? Player) ?: return